src/requirements.txt
beautifulsoup4
docker
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
flake8
funcy
matplotlib
numpy
pandas
pyqt5==5.15.2
pyqtchart==5.15.2
pyspellchecker
pytest
python-dateutil
requests
scikit-learn
spacy
tqdm
wheel
wordcloud
